# src/main/resources/static/api-docs/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/especialidades/{id}": {
      "get": {
        "tags": [
          "Especialidade"
        ],
        "summary": "Busca uma especialidade pelo ID",
        "operationId": "buscarEspecialidade",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Especialidade não encontrada"
          },
          "200": {
            "description": "Especialidade encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidade"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Especialidade"
        ],
        "summary": "Atualiza uma especialidade",
        "operationId": "atualizarEspecialidade",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Especialidade não encontrada"
          },
          "200": {
            "description": "Especialidade atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidade"
                }
              }
            }
          },
          "400": {
            "description": "Especialidade inválida"
          }
        }
      },
      "delete": {
        "tags": [
          "Especialidade"
        ],
        "summary": "Deleta uma especialidade",
        "operationId": "deletarEspecialidade",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Especialidade deletada",
            "content": {
              "application/json": {
              }
            }
          },
          "404": {
            "description": "Especialidade não encontrada"
          }
        }
      }
    },
...
    "/v1/medicos/{idMedico}/especialidades/{idEspecialidade}": {
      "post": {
        "tags": [
          "Médico"
        ],
        "summary": "Vincula uma especialidade a um médico",
        "operationId": "vincularEspecialidade",
        "parameters": [
          {
            "name": "idMedico",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Especialidade vinculada ao médico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida"
          },
          "404": {
            "description": "Médico ou especialidade não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Médico"
        ],
        "summary": "Desvincula uma especialidade de um médico",
        "operationId": "desvincularEspecialidade",
        "parameters": [
          {
            "name": "idMedico",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Requisição inválida"
          },
          "404": {
            "description": "Médico ou especialidade não encontrada"
          },
          "200": {
            "description": "Especialidade desvinculada do médico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          }
        }
      }
    },
    "/v1/especialidades": {
      "get": {
        "tags": [
          "Especialidade"
        ],
        "summary": "Busca todas as especialidades",
        "operationId": "buscarTodasEspecialidades",
        "responses": {
          "204": {
            "description": "Nenhuma especialidade encontrada"
          },
          "200": {
            "description": "Especialidades encontradas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidade"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Especialidade"
        ],
        "summary": "Cria uma especialidade",
        "operationId": "criarEspecialidade",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EspecialidadeDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Especialidade criada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Especialidade"
                }
              }
            }
          },
          "400": {
            "description": "Especialidade inválida"
          }
        }
      }
    },
...
    "/v1/medicos/especialidade/{idEspecialidade}": {
      "get": {
        "tags": [
          "Médico"
        ],
        "summary": "Busca médicos por especialidade",
        "operationId": "buscarMedicosPorEspecialidade",
        "parameters": [
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médicos encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "Nenhum médico encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EspecialidadeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          }
        }
      },
...
      "Especialidade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Medico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          },
          "crm": {
            "type": "string"
          },
          "especialidades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Especialidade"
            }
          }
        }
      },
...
    }
  }
}